// src/bot/common/enums/purchases.enum.ts
export enum Purchases_uz {
    CREATING_PURCHASE = "Xarid yaratish",
    PENDING_ORDERS = "Jarayondagi buyurtmalar",
    CONFIRMED_ORDERS = "Tastiqlangan buyurtmalar",
    IN_WAY_ORDERS = "Yo'ldagi buyurtmalar",
    COMPLETED_ORDERS = "Tugallangan buyurtmalar",
}


export enum Purchases_ru {
    CREATING_PURCHASE = "Создать заказ",
    PENDING_ORDERS = "Заказы в процессе",
    CONFIRMED_ORDERS = "Подтвержденные заказы",
    IN_WAY_ORDERS = "Заказы в пути",
    COMPLETED_ORDERS = "Завершенные заказы",
}